+++ Chapter.Net.WPF.Controls/ChapterTreeListView/ChapterTreeListViewItem.cs
@@ -1,5 +1,5 @@
 ﻿// -----------------------------------------------------------------------------------------------------------------
-// <copyright file="TreeListViewItem.cs" company="my-libraries">
+// <copyright file="ChapterTreeListViewItem.cs" company="my-libraries">
 //     Copyright (c) David Wendland. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------------------------------------------------
@@ -11,38 +11,38 @@ using System.Windows;
 namespace Chapter.Net.WPF.Controls
 {
     /// <summary>
-    ///     Represents a single entry in the <see cref="TreeListView" />.
+    ///     Represents a single entry in the <see cref="ChapterTreeListView" />.
     /// </summary>
-    public class TreeListViewItem : ChapterTreeViewItem
+    public class ChapterTreeListViewItem : ChapterTreeViewItem
     {
-        static TreeListViewItem()
+        static ChapterTreeListViewItem()
         {
-            DefaultStyleKeyProperty.OverrideMetadata(typeof(TreeListViewItem), new FrameworkPropertyMetadata(typeof(TreeListViewItem)));
+            DefaultStyleKeyProperty.OverrideMetadata(typeof(ChapterTreeListViewItem), new FrameworkPropertyMetadata(typeof(ChapterTreeListViewItem)));
         }
 
         /// <summary>
         ///     Returns the level of the current item in the tree.
         /// </summary>
-        public int Level => VisualTreeAssist.GetParentsUntilCount<TreeListViewItem, TreeListView>(this);
+        public int Level => VisualTreeAssist.GetParentsUntilCount<ChapterTreeListViewItem, ChapterTreeListView>(this);
 
         /// <summary>
-        ///     Generates a new child item container to hold in the <see cref="TreeListViewItem" />.
+        ///     Generates a new child item container to hold in the <see cref="ChapterTreeListViewItem" />.
         /// </summary>
         /// <returns>The generated child item container</returns>
         protected override DependencyObject GetContainerForItemOverride()
         {
-            return new TreeListViewItem();
+            return new ChapterTreeListViewItem();
         }
 
         /// <summary>
         ///     Checks if the item is already the correct item container. If not the <see cref="GetContainerForItemOverride" />
         ///     will be used to generate the right container.
         /// </summary>
-        /// <param name="item">The item to shown in the <see cref="TreeListViewItem" />.</param>
+        /// <param name="item">The item to shown in the <see cref="ChapterTreeListViewItem" />.</param>
         /// <returns>True if the item is the correct item container already.</returns>
         protected override bool IsItemItsOwnContainerOverride(object item)
         {
-            return item is TreeListViewItem;
+            return item is ChapterTreeListViewItem;
         }
     }
 }