+++ Demo/ChapterTreeListView/ChapterTreeListViewControl.xaml.cs
@@ -1,5 +1,5 @@
 ﻿// -----------------------------------------------------------------------------------------------------------------
-// <copyright file="TreeListViewControl.xaml.cs" company="my-libraries">
+// <copyright file="ChapterTreeListViewControl.xaml.cs" company="my-libraries">
 //     Copyright (c) David Wendland. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------------------------------------------------
@@ -11,11 +11,11 @@ using System.ComponentModel;
 
 namespace Demo;
 
-public partial class TreeListViewControl : INotifyPropertyChanged
+public partial class ChapterTreeListViewControl : INotifyPropertyChanged
 {
     private Element _selectedItem;
 
-    public TreeListViewControl()
+    public ChapterTreeListViewControl()
     {
         InitializeComponent();